+++ packages/adapter-begin/index.js
@@ -4,34 +4,8 @@ const { writeFileSync, readFileSync, existsSync } = require('fs');
 const { resolve, join, relative } = require('path');
 const parse = require('@architect/parser');
 const child_process = require('child_process');
-const { prerender } = require('@sveltejs/app-utils/renderer');
+const { prerender, generate_manifest_module } = require('@sveltejs/app-utils/renderer');
 const { copy } = require('@sveltejs/app-utils/files');
-
-function write_manifest(manifest) {
-	return `module.exports = {
-		layout: ${JSON.stringify(manifest.layout)},
-		error: ${JSON.stringify(manifest.error)},
-		components: ${JSON.stringify(manifest.components)},
-		pages: [
-			${manifest.pages
-				.map(
-					(page) =>
-						`{ pattern: ${page.pattern}, parts: ${JSON.stringify(page.parts)} }`
-				)
-				.join(',\n\t\t\t')}
-		],
-		server_routes: [
-			${manifest.server_routes
-				.map(
-					(route) =>
-						`{ name: '${route.name}', pattern: ${route.pattern}, file: '${
-							route.file
-						}', params: ${JSON.stringify(route.params)} }`
-				)
-				.join(',\n\t\t\t')}
-		]
-	};`.replace(/^\t/gm, '');
-}
 
 function parse_arc(arcPath) {
 	if (!existsSync(arcPath)) {
@@ -70,13 +44,13 @@ module.exports = async function builder({ dir, manifest, log }) {
 		join(static_directory, '_app')
 	);
 
-	// log.minor('Prerendering static pages...');
-	// await prerender({
-	//   dir,
-	//   out: static_directory,
-	//   manifest,
-	//   log
-	// });
+	log.minor('Prerendering static pages...');
+	await prerender({
+	  dir,
+	  out: static_directory,
+	  manifest,
+	  log
+	});
 
 	log.minor('Building lambda...' + lambda_directory);
 	copy(resolve(__dirname, 'src'), lambda_directory);