+++ packages/adapter-vercel/index.js
@@ -1,33 +1,7 @@
 const { writeFileSync, readFileSync } = require('fs');
 const { resolve, join } = require('path');
-const { prerender } = require('@sveltejs/app-utils/renderer');
+const { prerender, generate_manifest_module } = require('@sveltejs/app-utils/renderer');
 const { copy } = require('@sveltejs/app-utils/files');
-
-function write_manifest(manifest) {
-	return `module.exports = {
-		layout: ${JSON.stringify(manifest.layout)},
-		error: ${JSON.stringify(manifest.error)},
-		components: ${JSON.stringify(manifest.components)},
-		pages: [
-			${manifest.pages
-				.map(
-					page =>
-						`{ pattern: ${page.pattern}, parts: ${JSON.stringify(page.parts)} }`
-				)
-				.join(',\n\t\t\t')}
-		],
-		endpoints: [
-			${manifest.endpoints
-				.map(
-					route =>
-						`{ name: '${route.name}', pattern: ${route.pattern}, file: '${
-							route.file
-						}', params: ${JSON.stringify(route.params)} }`
-				)
-				.join(',\n\t\t\t')}
-		]
-	};`.replace(/^\t/gm, '');
-}
 
 module.exports = async function builder({ dir, manifest, log }) {
 	const lambda_directory = resolve('api');
@@ -41,7 +15,7 @@ module.exports = async function builder({ dir, manifest, log }) {
 	log.minor('Building lambda...');
 	copy(resolve(__dirname, 'src'), lambda_directory);
 	copy(join(resolve(dir), 'client.json'), join(server_directory, 'client.json'));
-	const written_manifest = write_manifest(manifest);
+	const written_manifest = generate_manifest_module(manifest);
 	const htmlPath = resolve('src', 'app.html');
 	const appHtml = readFileSync(htmlPath, 'utf-8');
 	writeFileSync(join(server_directory, 'manifest.js'), written_manifest);