+++ packages/app-utils/src/renderer/index.ts
@@ -1,2 +1,3 @@
 export { render } from './render';
 export { prerender } from './prerender';
+export { generate_manifest_module } from './manifest';