+++ packages/adapter-begin/CHANGELOG.md
@@ -1,6 +1,15 @@
 # @sveltejs/adapter-begin
 
+## 0.0.2
+
+### Patch Changes
+
+- 67eaeea: Move app-utils stuff into subpackages
+- Updated dependencies [67eaeea]
+  - @sveltejs/app-utils@0.0.16
+
 ## 0.0.1
+
 ### Patch Changes
 
 - Updated dependencies [a163000]